Linking up frontend and backend. Currently very messy, will need to reconfig models/repositories and how images are stored.

+++ src/main/java/com/github/thomaselliott/simplebattlemap/controller/TokenController.java
@@ -50,7 +50,7 @@ public class TokenController {
     }
 
     @RequestMapping(value = "/remove/{id}", method = RequestMethod.DELETE)
-    public void removeToken(@PathVariable Integer id) {
+    public void removeToken(@PathVariable Long id) {
         mapService.removeToken(id);
     }
 }
